+++ main.go
@@ -18,6 +18,7 @@ import (
 	"github.com/giantswarm/aws-admission-controller/v2/pkg/aws/awscontrolplane"
 	"github.com/giantswarm/aws-admission-controller/v2/pkg/aws/awsmachinedeployment"
 	"github.com/giantswarm/aws-admission-controller/v2/pkg/aws/g8scontrolplane"
+	"github.com/giantswarm/aws-admission-controller/v2/pkg/aws/machinedeployment"
 	"github.com/giantswarm/aws-admission-controller/v2/pkg/aws/networkpool"
 	"github.com/giantswarm/aws-admission-controller/v2/pkg/mutator"
 	"github.com/giantswarm/aws-admission-controller/v2/pkg/validator"
@@ -46,6 +47,11 @@ func main() {
 	}
 
 	g8scontrolplaneMutator, err := g8scontrolplane.NewMutator(config)
+	if err != nil {
+		panic(microerror.JSON(err))
+	}
+
+	machinedeploymentMutator, err := machinedeployment.NewMutator(config)
 	if err != nil {
 		panic(microerror.JSON(err))
 	}
@@ -82,6 +88,7 @@ func main() {
 	handler.Handle("/mutate/awsmachinedeployment", mutator.Handler(awsmachinedeploymentMutator))
 	handler.Handle("/mutate/awscontrolplane", mutator.Handler(awscontrolplaneMutator))
 	handler.Handle("/mutate/g8scontrolplane", mutator.Handler(g8scontrolplaneMutator))
+	handler.Handle("/mutate/machinedeployment", mutator.Handler(machinedeploymentMutator))
 	handler.Handle("/validate/awscluster", validator.Handler(awsclusterValidator))
 	handler.Handle("/validate/awscontrolplane", validator.Handler(awscontrolplaneValidator))
 	handler.Handle("/validate/awsmachinedeployment", validator.Handler(awsmachinedeploymentValidator))